+++ court-manager/src/App.js
@@ -1,13 +1,19 @@
-import logo from './logo.svg';
 import './App.css';
+import React, { useState } from 'react';
 import ImportButton from './components/ImportButton';
+import StatusTable from './components/StatusTable';
 
 function App() {
+  const [players, setPlayers] = useState(null);
   return (
     <div className='App'>
       <header className='App-header'>
         <h1>COURT MANAGER</h1>
-        <ImportButton shouldShowTestButton={true}></ImportButton>
+        <ImportButton
+          shouldShowTestButton={true}
+          setPlayers={setPlayers}
+        ></ImportButton>
+        <div>{players && <StatusTable players={players} />}</div>
       </header>
     </div>
   );
